11/28 학습 일지 및 대대적 수정 초안
UML관련 리펙토리링 예정

기존 코드 구성 변경 및 프로젝트 전면 수정
1. 리소스 매니저 추가(게임에 사용되는 BufferredImage를 받아 동기적으로 사용 가능하게할 예정)
2. 게임 메니저의 중추 인테페이스 역할 강화
3. 게임 루프의 싱글턴 변경
이후 수기로 목록작성 및 수정 후 UML 예상 작성

포인트: 코드 디자인 패턴을 최대한 따라서 객체 지향 설계를 추구하자

+++ LocalPVP2PShooting/src/core/Camera.java
@@ -4,17 +4,17 @@ import view.GameWindow;
 
 public class Camera {
 
-	private static final int MAP_WIDTH = MapManager.MAP_WIDTH;
-	private static final int MAP_HEIGHT = MapManager.MAP_HEIGHT;
+	private static final double MAP_WIDTH = MapManager.MAP_WIDTH;
+	private static final double MAP_HEIGHT = MapManager.MAP_HEIGHT;
 
 	private double playerWorldX;
 	private double playerWorldY;
 	private double cameraWorldX = 0.0;
 	private double cameraWorldY = 0.0;
 	private GameWindow window;
-	private static final int DEAD_ZONE_WIDTH = 1;
-	private static final int DEAD_ZONE_HEIGHT = 1;
-	private static final int FIXED_RANDER_SIZE = 5000;
+	private static final double DEAD_ZONE_WIDTH = 1;
+	private static final double DEAD_ZONE_HEIGHT = 1;
+	private static final double FIXED_RANDER_SIZE = 5000;
 
 	public Camera() {
 	}
@@ -29,13 +29,13 @@ public class Camera {
 	}
 
 	public void updateCamera() {
-		int viewPortWidth = window.getWidth();
-		int viewPortHeight = window.getHeight();
+		double viewPortWidth = window.getWidth();
+		double viewPortHeight = window.getHeight();
 
-		int deadZoneMinX = ((viewPortWidth - DEAD_ZONE_WIDTH) / 2);
-		int deadZoneMaxX = (deadZoneMinX + DEAD_ZONE_WIDTH);
-		int deadZoneMinY = ((viewPortHeight - DEAD_ZONE_HEIGHT) / 2);
-		int deadZoneMaxY = (deadZoneMinY + DEAD_ZONE_HEIGHT);
+		double deadZoneMinX = ((viewPortWidth - DEAD_ZONE_WIDTH) / 2);
+		double deadZoneMaxX = (deadZoneMinX + DEAD_ZONE_WIDTH);
+		double deadZoneMinY = ((viewPortHeight - DEAD_ZONE_HEIGHT) / 2);
+		double deadZoneMaxY = (deadZoneMinY + DEAD_ZONE_HEIGHT);
 
 		double playerDeadZoneRefX = playerWorldX - cameraWorldX;
 		double playerDeadZoneRefY = playerWorldY - cameraWorldY;
@@ -59,7 +59,7 @@ public class Camera {
 		cameraWorldY = Math.max(0.0, Math.min(cameraWorldY, maxCameraWorldY));
 	}
 
-	public int getCameraSize() {
+	public double getCameraSize() {
 		return FIXED_RANDER_SIZE;
 	}
 
@@ -71,11 +71,11 @@ public class Camera {
 		return playerWorldY;
 	}
 
-	public int getCameraWorldX() {
-		return (int) Math.round(cameraWorldX);
+	public double getCameraWorldX() {
+		return (double) Math.round(cameraWorldX);
 	}
 
-	public int getCameraWorldY() {
-		return (int) Math.round(cameraWorldY);
+	public double getCameraWorldY() {
+		return (double) Math.round(cameraWorldY);
 	}
 }
